Begin writing C++, Python and Julia code
Problems a-b) are written in C++ and Julia, both running at the same speed. The
C++ results are saved as raw ascii and read by Python which in turns plots them.
Since Julia is a gazillion times better, the plotting is done internally in
Julia.

Both of the codes have a bug causing the graphs to be of correct shape, but
wrong magnitude, where the magnitude is dependent on the size of the matrix.