Improve incoporation of Python with LaTeX
A tagging system is now available, allowing arbitrary messages
to be passed from a python script to PythonTex. Any data to be
handled in the latex file is tagged using latextools.tag(identifier, data).
On the recieving end, the latex file calls the analysis script using
subprocessing and a pipe, where the data is decoded to utf-8 and
untagged using latextools.untag_all(data). The untagged data is
placed into a directory of the form {identifier: data, ...} for easy access.