Fix LU boundary conditions
The boundary conditions for LU-decomposition can easily be set by
extending the matrix by two rows where there is a 1 at the diagonal and 0
everywhere else, while at the same time setting btilde(0)=u(0) and
btilde(end) = u(end).

If no boundary conditions are specified, Dirichlet boundary conditions are
assumed.

The same methods was attempted on the general solver, but to no luck.
I assume this is because the indexation is wrong.